moved some tests from nitrogen
