+++ RUFAS/routines/field/crop/nutrient_uptake.py
@@ -38,6 +38,7 @@ class NutrientUptake:
         layer_demand = [sum(layer_delta[:i]) for i in range(len(layer_delta))]
         return [max(val, 0) for val in layer_demand]
 
+    # tested
     @staticmethod
     def determine_potential_nutrient_uptake(
         demand: float,
@@ -153,6 +154,7 @@ class NutrientUptake:
         )
         self.crop_data.inaccessible_soil_layers = max(len(depths) - self.crop_data.accessible_soil_layers, 0)
 
+    # tested
     @staticmethod
     def _determine_deepest_accessible_layer(root_depth: float, layer_bounds: list[float]) -> int:
         """
@@ -218,8 +220,9 @@ class NutrientUptake:
             A trimmed list with an element for each soil layer that is accessible to the plant's roots.
 
         """
-        return layer_list[0 : self.crop_data.accessible_soil_layers]
+        return layer_list[0: self.crop_data.accessible_soil_layers]
 
+    # tested
     @classmethod
     def determine_layer_nutrient_uptake_potential(
         cls,
@@ -288,39 +291,40 @@ class NutrientUptake:
             raise ValueError("multiple soil boundaries cannot have the same depths. Remove the redundant layer?")
 
         boundary_nutrient = [
-            cls._determine_nitrogen_uptake_to_depth(total_demand, x, root_depth, nutrient_distribution_parameter)
+            cls._determine_nutrient_uptake_to_depth(total_demand, x, root_depth, nutrient_distribution_parameter)
             for x in layer_bounds
         ]
         boundary_nutrient.insert(0, 0)
         layer_nutrient = [below - above for below, above in zip(boundary_nutrient[1:], boundary_nutrient)]
         return layer_nutrient
 
+    # tested
     @staticmethod
-    def _determine_nitrogen_uptake_to_depth(
+    def _determine_nutrient_uptake_to_depth(
         demand: float,
         depth: float,
         root_depth: float,
         nutrient_distribution_parameter: float,
     ) -> float:
         """
-        Calculates the potential nitrogen uptake from the soil surface to a specified depth.
+        Calculates the potential nutrient uptake from the soil surface to a specified depth.
 
         Parameters
         ----------
         demand : float
-            The current nitrogen demand of the plant (kg/ha).
+            The current nutrient demand of the plant (kg/ha).
         depth : float
             The depth (in the same units as root_depth, typically centimeters or meters) to which nitrogen uptake is
             calculated (mm).
         root_depth : float
             The current depth of the plant's roots (mm).
         nutrient_distribution_parameter : float
-            The nitrogen uptake distribution parameter affecting how uptake is allocated with depth.
+            The nutrient uptake distribution parameter affecting how uptake is allocated with depth.
 
         Returns
         -------
         float
-            The potential amount of nitrogen that can be taken up from the soil surface to the specified depth (kg/ha).
+            The potential amount of nutrient that can be taken up from the soil surface to the specified depth (kg/ha).
 
         References
         ----------
@@ -329,17 +333,17 @@ class NutrientUptake:
         """
         info_map = {
             "class": NutrientUptake.__class__.__name__,
-            "function": NutrientUptake._determine_nitrogen_uptake_to_depth.__name__,
+            "function": NutrientUptake._determine_nutrient_uptake_to_depth.__name__,
         }
         om = OutputManager()
         if nutrient_distribution_parameter == 0:
             om.add_error(
-                "Invalid nitrogen_distribution_parameter.",
-                "Received invalid value 0 for nitrogen_distribution_parameter. 0 nitrogen_distribution_parameter"
+                "Invalid nutrient_distribution_parameter.",
+                "Received invalid value 0 for nutrient_distribution_parameter. 0 nutrient_distribution_parameter"
                 " will lead to exp(0) calculation.",
                 info_map,
             )
-            raise ValueError("nitrogen_distribution_parameter cannot equal 0")
+            raise ValueError("nutrient_distribution_parameter cannot equal 0")
         if root_depth <= 0:
             return 0
         else:
@@ -386,6 +390,7 @@ class NutrientUptake:
         if self.crop_data.inaccessible_soil_layers > 0:
             actual_nutrient_uptakes += [0] * self.crop_data.inaccessible_soil_layers
 
+    # tested
     @classmethod
     def determine_nutrient_shape_parameters(
         cls,
@@ -474,6 +479,7 @@ class NutrientUptake:
         s1 = log_term + s2 * half_mature_heat_fraction
         return [s1, s2]
 
+    # tested
     @staticmethod
     def _determine_shape_log(
         heat_fraction: float,
@@ -607,6 +613,7 @@ class NutrientUptake:
             )
         return log((heat_fraction / denominator) - heat_fraction)
 
+    # tested
     @staticmethod
     def determine_optimal_nutrient_fraction(
         heat_fraction: float,
@@ -646,6 +653,7 @@ class NutrientUptake:
         brackets = 1 - (heat_fraction / (heat_fraction + e_term))
         return (ndiff * brackets) + mature_nutrient_fraction
 
+    # tested
     @staticmethod
     def determine_optimal_nutrient(fraction: float, whole: float) -> float:
         """
